Competition test

+++ 2018 Competition Robot/src/org/usfirst/frc/team1732/robot/sensors/encoders/TalonEncoder.java
@@ -34,7 +34,7 @@ public class TalonEncoder extends EncoderBase {
 	}
 
 	@Override
-	public double getPulses() {
+	public int getPulses() {
 		return talon.getSelectedSensorPosition(0);
 	}
